Refactor UserProfilePage to update component imports and remove obsolete OrderSummaryCards and RecentOrdersList components, streamlining the file structure.
